requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-filter==24.2
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
PyJWT==2.8.0
bleach==6.1.0
pillow==10.3.0
environs==11.0.0
redis==5.0.5
psycopg2-binary==2.9.9
minio==7.2.7
boto3==1.35.23
